add readme solid
