+++ solid/readme.md
@@ -1,0 +1,4 @@
+# This repository for learn solid
+- Source:
+    + https://www.digitalocean.com/community/conceptual_articles/s-o-l-i-d-the-first-five-principles-of-object-oriented-design#open-closed-principle
+    + https://www.youtube.com/watch?v=f13Bff-wiJ4&list=PLoaAbmGPgTSMpPS7L1vFDRsPCFvncUc_D&index=8
